# storage-service/storage-client/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: RustFS Storage Service API
  version: 1.0.0
  description: File storage service backed by RustFS (S3-compatible)

paths:
  /v1/files:
    post:
      operationId: uploadFile
      summary: Upload a file
      tags: [files]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                tags:
                  type: string
                  description: JSON object of key-value tag pairs
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'

  /v1/files/search:
    get:
      operationId: searchFiles
      summary: Search files by metadata
      tags: [files]
      parameters:
        - name: prefix
          in: query
          schema:
            type: string
        - name: mimeType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of matching files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileResponse'

  /v1/files/{fileId}:
    get:
      operationId: getFile
      summary: Stream file content
      tags: [files]
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteFile
      summary: Delete a file
      tags: [files]
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '204':
          description: File deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/files/{fileId}/presigned-url:
    get:
      operationId: getPresignedUrl
      summary: Get a temporary presigned download URL (valid 1 hour)
      tags: [files]
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/files/{fileId}/metadata:
    patch:
      operationId: updateMetadata
      summary: Update custom S3 tags on a file
      tags: [files]
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        '200':
          description: Metadata updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    NotFound:
      description: File not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    FileUploadResponse:
      type: object
      properties:
        fileId:
          type: string
          format: uuid
        objectKey:
          type: string

    FileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        size:
          type: integer
          format: int64
        mimeType:
          type: string
        objectKey:
          type: string
        createdAt:
          type: string
          format: date-time

    PresignedUrlResponse:
      type: object
      properties:
        url:
          type: string
        expiresAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time